cloud-computing-project/requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
psutil==5.8.0
requests==2.26.0    